messages added

+++ app/Modules/Friends/Repositories/FriendsRepository.php
@@ -15,16 +15,17 @@ class FriendsRepository extends Repository implements FriendsRepositoryInterface
     {
         switch($status) {
             case 1:
-                return $this->model->where( 'friend_id', $id )->delete();
+                return $this->model->where( 'friend_id', $id )->where( 'user_id', \Auth::user()->id )
+                    ->orWhere( 'user_id', $id )->where( 'friend_id', \Auth::user()->id )->delete();
                 break;
             case 2:
-                return $this->model->where( 'friend_id', $id )->delete();
+                return $this->model->where( 'friend_id', $id )->where( 'user_id', \Auth::user()->id )->delete();
 
             case 3:
-                return $this->model->where( 'user_id', $id )->where( 'friend_id', \Auth::user()->id )->update(['friendship' => 1]);;
+                return $this->model->where( 'user_id', $id )->where( 'friend_id', \Auth::user()->id )->update(['friendship' => 1]);
 
             case 4:
-                return $this->model->where( 'friend_id', $id )->delete();
+                return $this->model->where( 'user_id', $id )->where( 'friend_id', \Auth::user()->id )->delete();
 
             default:
                 return $this->model->create([
@@ -39,7 +40,10 @@ class FriendsRepository extends Repository implements FriendsRepositoryInterface
 
     public function friendshipStatus( $id ){
         return $this->model->where( 'user_id', \Auth::user()->id )
-            ->where( 'friend_id', $id )->get();
+            ->where( 'friend_id', $id )
+            ->orWhere( 'friend_id', \Auth::user()->id )
+            ->where( 'user_id', $id )
+            ->get();
     }
 
     public function userFriends( $id )
